+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Regular release sync fix - this time it was to do with JSON serialisation [[#1072](https://github.com/nf-core/tools/pull/1072)]
 * Fixed bug in schema validation that ignores upper/lower-case typos in parameters [[#1087](https://github.com/nf-core/tools/issues/1087)]
 * Bugfix: Download should use path relative to workflow for configs
+* Added lint check for valid default parameterse in `nextflow.config` [[#992](https://github.com/nf-core/tools/issues/992)]
 
 ### Modules